Fall back to Ember for auth
Ref https://linear.app/ghost/issue/BER-2917/fall-back-to-ember-for-authentication

If we can't fetch the current user, fall back to displaying the Ember app so
that it can handle authentication for us. Once the user navigates away from any
known auth route, refetch the user so that we're up to date on the React side.

This is a temporary solution to handle auth until we bridge the Ember state 
properly or re-implement auth on the React side.
