+++ apps/admin/src/routes.tsx
@@ -4,7 +4,7 @@ import GlobalDataProvider from "@tryghost/stats/src/providers/GlobalDataProvider
 import {FeatureFlagsProvider} from "@tryghost/activitypub/src/lib/feature-flags";
 import { routes as statsRoutes, APP_ROUTE_PREFIX as statsAppRoutePrefix } from "@tryghost/stats/src/routes";
 import { routes as activityPubRoutes, APP_ROUTE_PREFIX as activityPubAppRoutePrefix } from "@tryghost/activitypub/src/routes";
-import EmberFallback from "./ember-bridge/EmberFallback";
+import { EmberFallback } from "./ember-bridge";
 
 export const routes: RouteObject[] = [
     {
